+++ examples/dicts/upload.py
@@ -1,53 +1,33 @@
-from pydantic import BaseModel, Field
-
 from pipeline import Pipeline, Variable, pipe
-from pipeline.cloud.compute_requirements import Accelerator
-from pipeline.cloud.environments import create_environment
 from pipeline.cloud.pipelines import upload_pipeline
+from pipeline.objects.graph import InputField, InputSchema
 
 
-class MyKwargs(BaseModel):
-    a: int = Field(10, lt=20, gte=1)
-    b: str = Field("hello", min_length=2, max_length=10)
+class MyInputSchema(InputSchema):
+    in_1: int = InputField(lt=5, gt=-5, description="kwarg 1", title="kwarg_1")
+    in_2: int | None = InputField(
+        default=0, lt=5, gt=-5, description="kwarg 1", title="kwarg_1"
+    )
 
 
 @pipe
-def do_something(kwargs: dict) -> str:
-    return str(kwargs)
+def my_func(in_1: int, other_schema: MyInputSchema) -> int:
+    return in_1 + other_schema.in_2 + other_schema.in_1
 
 
 with Pipeline() as builder:
-    input_var = Variable(
-        dict,
-        dict_schema=MyKwargs,
-    )
+    var_1 = Variable(int, lt=10, ge=0)
+    var_2 = Variable(MyInputSchema)
 
-    o = do_something(input_var)
+    output = my_func(var_1, var_2)
+    builder.output(output)
 
-    builder.output(o)
+my_pl = builder.get_pipeline()
 
-pl = builder.get_pipeline()
-
-
-env_id = create_environment(
-    name="numpy",
-    python_requirements=["numpy==1.24.3"],
-    allow_existing=True,
-)
-
-print(f"New environment ID = {env_id}")
-
-result = upload_pipeline(
-    pl,
-    "paulh/dict-test:test",
-    environment_id_or_name="numpy",
+rm_pipeline = upload_pipeline(
+    my_pl,
+    "schema-demo",
+    "numpy",
     minimum_cache_number=1,
-    required_gpu_vram_mb=None,
-    accelerators=[
-        Accelerator.cpu,
-    ],
 )
-
-
-pipeline_id = result.id
-print(f"New pipeline ID = {pipeline_id}")
+print(rm_pipeline.id)
